backend/requirements.txt
# FastAPI e servidor
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Validação de dados
pydantic==2.5.0
pydantic-settings==2.1.0

# Base de dados
motor==3.3.2
pymongo==4.6.0

# Groq API (alternativa gratuita à OpenAI)
groq==0.4.2

# Variáveis de ambiente
python-dotenv==1.0.0

# Processamento de imagens (para futuras funcionalidades)
pillow==10.1.0

# Utilitários
python-multipart==0.0.6
httpx==0.25.2

# Desenvolvimento e testes
pytest==7.4.3
pytest-asyncio==0.21.1
